Fix ClipperEngine.cpp narrowing conversion error on MSVC
MSVC is strict about narrowing conversions in brace initialization.
Cast edge[0] to size_t explicitly when inserting into the map.

+++ tools/Boolean/Clipper/ClipperEngine.cpp
@@ -35,7 +35,7 @@ namespace ClipperEngineHelper {
             if (itr != adjacencies.end()) {
                 itr->second.push_back(edge[1]);
             } else {
-                adjacencies.insert({edge[0], {edge[1]}});
+                adjacencies.insert({static_cast<size_t>(edge[0]), {edge[1]}});
             }
         }
